requirements.txt
streamlit
opencv-python
numpy
fer
Pillow
moviepy
